move in model call specific utils from server utils
